Moved the TodoContainer.jsx file and its associated CSS module file (TodoContainer.module.css) to the components/ directory. Did so according to the stretch goal directions outlined in the bullet points underneath the 'Improve Project Structure' header, in the instructions for completing the assignment for lesson 4.1, which are found in the CTD Learns App.
